+++ gammapy/data/data_store.py
@@ -209,6 +209,25 @@ class DataStore(object):
         else:
             raise ValueError('Invalid filetype.')
 
+    def load_all(self, filetype):
+        """Load a given file type for all observations
+
+        Parameters
+        ----------
+        filetype : {'events', 'aeff', 'edisp', 'psf', 'background'}
+            Type of file.
+
+        Returns
+        -------
+        list : python list of object
+            Object depends on type, e.g. for `events` it is a list of `~gammapy.data.EventList`.
+        """
+        data_lists = []
+        for obs_id in self.obs_table['OBS_ID']:
+            data_list = self.load(obs_id, filetype)
+            data_lists.append(data_list)
+        return data_lists
+        
     def check_integrity(self, logger):
         """Check integrity, i.e. whether index table and files match.
         """