+++ gammapy/scripts/tests/test_image_fit.py
@@ -2,14 +2,29 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 import json
 from numpy.testing.utils import assert_allclose
+from astropy.stats import gaussian_sigma_to_fwhm
+from astropy.tests.helper import pytest
 from ...utils.testing import requires_dependency, requires_data
 from ...datasets import load_poisson_stats_image
 from ..image_fit import image_fit
 
 
+EXPECTED = ([9.016526, 99.865985, 100.147877, 1010.824189],
+            [4 * gaussian_sigma_to_fwhm, 100, 100, 1E3],
+            [5 * gaussian_sigma_to_fwhm, 100, 100, 1E3])
+RTOL = (1E-5, 1E-3, 1E-3)
+PSF = (True, True, False)
+DATA = (True, False, False)
+
 @requires_dependency('sherpa')
 @requires_data('gammapy-extra')
-def test_sherpa_like(tmpdir):
+@pytest.mark.parametrize('expected, rtol, psf, data',
+                          zip(EXPECTED, RTOL, PSF, DATA))
+def test_sherpa_like(tmpdir, expected, rtol, psf, data):
+    """
+    Fit Poisson stats image test data.
+    """
+
     # load test data
     filenames = load_poisson_stats_image(extra_info=True, return_filenames=True)
     outfile = tmpdir / 'test_sherpa_like.json'
@@ -19,30 +34,36 @@ def test_sherpa_like(tmpdir):
     sources_data['gaussian'] = {'ampl': 1E3,
                                 'xpos': 99,
                                 'ypos': 99,
-                                'fwhm': 4 * 2.3548}
+                                'fwhm': 4 * gaussian_sigma_to_fwhm}
 
     filename = tmpdir / 'test_sherpa_like_sources.json'
     with filename.open('w') as fh:
         json.dump(sources_data, fh)
 
     # set up args
-    args = {'counts': str(filenames['counts']),
-            'exposure': str(filenames['exposure']),
+    args = {'exposure': str(filenames['exposure']),
             'background': str(filenames['background']),
-            'psf': filenames['psf'],
             'sources': str(filename),
             'roi': None,
             'outfile': str(outfile)}
+
+    if data:
+        args['counts'] = str(filenames['counts'])
+    else:
+        args['counts'] = str(filenames['model'])
+    if psf:
+        args['psf'] = filenames['psf']
+    else:
+        args['psf'] = None
+
     image_fit(**args)
 
     with outfile.open() as fh:
         data = json.load(fh)
 
-    # Note: the reference results here changed once and
-    # we didn't track down why at the time:
-    # See https://github.com/gammapy/gammapy/issues/349
-    # old: [  9.016334  ,  99.365574   ,  99.647234   , 10.97365    ]
-    # new: [ 10.7427035 ,  98.16618776 ,  98.45487028 ,  7.73529899 ]
+    # This recovers the values from the test dataset documented here:
+    # https://github.com/gammapy/gammapy-extra/tree/master/
+    # test_datasets/unbundled/poisson_stats_image#data
     actual = data['fit']['parvals']
-    expected = [10.7427035, 98.16618776, 98.45487028, 7.73529899]
-    assert_allclose(actual, expected)
+    assert_allclose(actual, expected, rtol=rtol)
+