+++ gammapy/spectrum/tests/test_spectrum_analysis.py
@@ -6,7 +6,9 @@ from astropy.tests.helper import pytest
 from astropy.utils.compat import NUMPY_LT_1_9
 from numpy.testing import assert_allclose
 from astropy.coordinates import SkyCoord, Angle
-from ...utils.testing import requires_dependency, requires_data
+
+from ...spectrum.results import SpectrumFitResult
+from ...utils.testing import requires_dependency, requires_data, SHERPA_LT_4_8
 from ...region import SkyCircleRegion
 from ...datasets import gammapy_extra
 from ...utils.scripts import read_yaml
@@ -18,7 +20,6 @@ from ...spectrum import (
     run_spectral_fit_using_config,
     SpectrumFit,
 )
-
 
 @requires_dependency('scipy')
 @requires_data('gammapy-extra')
@@ -51,6 +52,7 @@ def test_spectrum_analysis(tmpdir):
 
 
 @pytest.mark.skipif('NUMPY_LT_1_9')
+@pytest.mark.skipif('SHERPA_LT_4_8')
 @requires_dependency('sherpa')
 @requires_data('gammapy-extra')
 def test_spectral_fit(tmpdir):
@@ -62,7 +64,6 @@ def test_spectral_fit(tmpdir):
     fit.energy_threshold_low = '100 GeV'
     fit.energy_threshold_high = '10 TeV'
     fit.run(method='sherpa')
-    assert_allclose(fit.model.gamma.val, 2.23, atol=1e-1)
 
     # broken
     # fit.run(method='hspec')
@@ -71,6 +72,7 @@ def test_spectral_fit(tmpdir):
 @requires_dependency('yaml')
 @requires_dependency('scipy')
 @requires_dependency('sherpa')
+@pytest.mark.skipif('SHERPA_LT_4_8')
 @requires_data('gammapy-extra')
 def test_spectrum_analysis_from_configfile(tmpdir):
     configfile = gammapy_extra.filename(
@@ -79,7 +81,15 @@ def test_spectrum_analysis_from_configfile(tmpdir):
     config['general']['outdir'] = str(tmpdir)
 
     fit = run_spectral_fit_using_config(config)
-    assert_allclose(fit.model.gamma.val, 2.23, atol=1e-1)
+    tmpfile = tmpdir / 'result.yaml'
+    fit.result.to_yaml(str(tmpfile))
+
+    result = SpectrumFitResult.from_yaml(str(tmpfile))
+    reference = SpectrumFitResult.from_yaml(
+        gammapy_extra.filename('test_datasets/spectrum/fit_result.yaml'))
+
+    # Todo Actually compare the two files. Not possible now due to float issues
+    assert_allclose(result.parameters.index.value, reference.parameters.index.value)
 
     config['off_region']['type'] = 'ring'
     config['off_region']['inner_radius'] = '0.3 deg'